e2e tests: further split the e2e/client pkg by files
